refactor: move o adapter simples pra diretorio de adapters
